gpu/recompiler: lift the shared recompile analyses into a header
Step one of splitting rdna2_to_spirv.cpp (4,763 lines) per shader stage. 26 regions move:
the branch and exec-mask analyses and the specializations every stage's recompile path
calls. The set is computed rather than chosen -- split_file.py named the definitions a
per-stage split would strand, and promote_internal closed that under its own callees.

rdna2_to_spirv.cpp -> 3608 lines; rdna2_recompile_shared.hpp 1196. Builds clean.

Promoted in ONE run: a second promotion naming the same header would replace it rather
than extend it (#3523).
